+++ base/src/testFixtures/java/eu/fraho/spring/securityJwt/base/it/spring/TestApiApplication.java
@@ -6,15 +6,13 @@
  */
 package eu.fraho.spring.securityJwt.base.it.spring;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication(scanBasePackages = {"eu.fraho.spring.securityJwt"})
+@Slf4j
 public class TestApiApplication {
-    private static final Logger log = LoggerFactory.getLogger(TestApiApplication.class);
-
     public static void main(String[] args) {
         log.info("Starting spring context");
         SpringApplication.run(TestApiApplication.class, args);