+++ hibernate/src/test/java/eu/fraho/spring/securityJwt/hibernate/AuthControllerHibernateTest.java
@@ -7,14 +7,14 @@
 package eu.fraho.spring.securityJwt.hibernate;
 
 import eu.fraho.spring.securityJwt.base.it.AbstractAuthControllerWithRefreshTest;
-import eu.fraho.spring.securityJwt.hibernate.spring.TestHibernateApiApplication;
+import eu.fraho.spring.securityJwt.base.it.spring.TestApiApplication;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.transaction.annotation.Transactional;
 
-@SpringBootTest(properties = "spring.config.location=classpath:hibernate-test.yaml", classes = TestHibernateApiApplication.class)
+@SpringBootTest(properties = "spring.config.location=classpath:hibernate-test.yaml", classes = TestApiApplication.class)
+@EntityScan(basePackages = {"eu.fraho.spring.securityJwt.hibernate"})
 @ExtendWith(SpringExtension.class)
-@Transactional
 public class AuthControllerHibernateTest extends AbstractAuthControllerWithRefreshTest {
 }